pyproject.toml
[project]
name = "agent-toolkit"
version = "0.1.0"
description = "AI-powered web research agent using Apify + Contextual AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "contextual-client",
    "apify-client",
    "streamlit",
    "python-dotenv",
    "mcp>=1.26.0",
]
